requirements.txt
flask
netmiko
requests
python-dotenv
PyYAML
tenacity
